Split app/ai_coach.py and app/api.py monoliths (P3-1)
Mechanical, zero-behavior-change refactor: app/ai_coach.py (2,586 lines)
becomes an app/coach/ package (context, providers, jobs, plans, chat), and
app/api.py (2,206 lines, 53 routes) becomes app/routers/ split by domain
(activities, daily, calendar, plan, races, chat, settings, trends, export).

Both app/ai_coach.py and app/api.py are now thin re-export shims so existing
imports keep working. A handful of call sites inside app/coach/* (db_session
opens, provider/job dispatch in providers.py/jobs.py/plans.py) route through
the app.ai_coach shim at call time rather than as bare intra-package names,
since the test suite monkeypatches these on app.ai_coach and a bare call
would otherwise resolve against the wrong module's globals.

All 795 tests pass; no test files were modified.
