+++ app/routers/_shared.py
@@ -1,0 +1,25 @@
+"""Small helpers shared across two or more domain routers."""
+from datetime import date, datetime
+
+from app import adherence as adherence_mod
+from app.models import GarminCalendarEvent
+from app.schemas import CalendarEventResponse
+
+
+def _parse_date(s: str | None) -> date:
+    if not s:
+        return date.today()
+    try:
+        return datetime.strptime(s, "%Y-%m-%d").date()
+    except ValueError:
+        return date.today()
+
+
+def _enrich_event_with_steps(event: GarminCalendarEvent) -> CalendarEventResponse:
+    """Convert a GarminCalendarEvent to CalendarEventResponse with parsed workout steps."""
+    resp = CalendarEventResponse.model_validate(event)
+    if event.event_type == "workout" and event.raw_json:
+        steps = adherence_mod.parse_workout_steps(event.raw_json)
+        if steps:
+            resp.workout_steps = steps
+    return resp
