+++ app/routers/calendar.py
@@ -1,0 +1,173 @@
+"""Calendar month/week views and calendar event detail."""
+import calendar as cal_mod
+from datetime import date, datetime, timedelta
+
+from fastapi import APIRouter, Depends, HTTPException, Query
+from sqlalchemy.orm import Session
+
+from app.auth import get_current_user
+from app.database import get_db
+from app.models import Activity, AthleteProfile, GarminCalendarEvent, User
+from app.schemas import ActivitySummary, CalendarDay, CalendarEventResponse, FuellingGuidance
+from app import nutrition as nutrition_mod
+from app.routers._shared import _enrich_event_with_steps, _parse_date
+
+router = APIRouter()
+
+
+@router.get("/calendar", response_model=list[CalendarDay])
+def api_calendar_month(
+    month: str = Query(None),
+    db: Session = Depends(get_db),
+    current_user: User = Depends(get_current_user),
+):
+    if month:
+        try:
+            view_date = datetime.strptime(month, "%Y-%m").date()
+        except ValueError:
+            view_date = date.today().replace(day=1)
+    else:
+        view_date = date.today().replace(day=1)
+
+    if view_date.month == 12:
+        next_month = view_date.replace(year=view_date.year + 1, month=1)
+    else:
+        next_month = view_date.replace(month=view_date.month + 1)
+
+    month_activities = (
+        db.query(Activity)
+        .filter(
+            Activity.user_id == current_user.id,
+            Activity.started_at >= datetime.combine(view_date, datetime.min.time()),
+            Activity.started_at < datetime.combine(next_month, datetime.min.time()),
+        )
+        .order_by(Activity.started_at.asc())
+        .all()
+    )
+
+    activities_by_date: dict[date, list] = {}
+    for a in month_activities:
+        if a.started_at:
+            d = a.started_at.date()
+            activities_by_date.setdefault(d, []).append(a)
+
+    month_events = (
+        db.query(GarminCalendarEvent)
+        .filter(
+            GarminCalendarEvent.user_id == current_user.id,
+            GarminCalendarEvent.date >= view_date,
+            GarminCalendarEvent.date < next_month,
+        )
+        .order_by(GarminCalendarEvent.date.asc())
+        .all()
+    )
+    events_by_date: dict[date, list] = {}
+    for e in month_events:
+        events_by_date.setdefault(e.date, []).append(e)
+
+    # Build all days of the month
+    cal = cal_mod.Calendar(firstweekday=0)
+    result = []
+    for day_num, _ in cal.itermonthdays2(view_date.year, view_date.month):
+        if day_num == 0:
+            continue
+        d = date(view_date.year, view_date.month, day_num)
+        result.append(CalendarDay(
+            date=d,
+            activities=[ActivitySummary.model_validate(a) for a in activities_by_date.get(d, [])],
+            events=[_enrich_event_with_steps(e) for e in events_by_date.get(d, [])],
+        ))
+    return result
+
+
+@router.get("/calendar/week", response_model=list[CalendarDay])
+def api_calendar_week(
+    date_str: str = Query(None, alias="date"),
+    db: Session = Depends(get_db),
+    current_user: User = Depends(get_current_user),
+):
+    target = _parse_date(date_str) if date_str else date.today()
+    # Monday of the week
+    week_start = target - timedelta(days=target.weekday())
+    week_end = week_start + timedelta(days=7)
+
+    activities = (
+        db.query(Activity)
+        .filter(
+            Activity.user_id == current_user.id,
+            Activity.started_at >= datetime.combine(week_start, datetime.min.time()),
+            Activity.started_at < datetime.combine(week_end, datetime.min.time()),
+        )
+        .order_by(Activity.started_at.asc())
+        .all()
+    )
+    activities_by_date: dict[date, list] = {}
+    for a in activities:
+        if a.started_at:
+            d = a.started_at.date()
+            activities_by_date.setdefault(d, []).append(a)
+
+    events = (
+        db.query(GarminCalendarEvent)
+        .filter(
+            GarminCalendarEvent.user_id == current_user.id,
+            GarminCalendarEvent.date >= week_start,
+            GarminCalendarEvent.date < week_end,
+        )
+        .all()
+    )
+    events_by_date: dict[date, list] = {}
+    for e in events:
+        events_by_date.setdefault(e.date, []).append(e)
+
+    result = []
+    for i in range(7):
+        d = week_start + timedelta(days=i)
+        result.append(CalendarDay(
+            date=d,
+            activities=[ActivitySummary.model_validate(a) for a in activities_by_date.get(d, [])],
+            events=[_enrich_event_with_steps(e) for e in events_by_date.get(d, [])],
+        ))
+    return result
+
+
+@router.get("/calendar-events/{event_id}", response_model=CalendarEventResponse)
+def api_calendar_event_detail(
+    event_id: int,
+    db: Session = Depends(get_db),
+    current_user: User = Depends(get_current_user),
+):
+    event = (
+        db.query(GarminCalendarEvent)
+        .filter(GarminCalendarEvent.id == event_id, GarminCalendarEvent.user_id == current_user.id)
+        .first()
+    )
+    if not event:
+        raise HTTPException(status_code=404, detail="Calendar event not found")
+    resp = _enrich_event_with_steps(event)
+
+    if event.event_type == "race":
+        duration_sec = (
+            event.goal_time_sec
+            or event.projected_race_time_sec
+            or event.predicted_race_time_sec
+        )
+        if duration_sec:
+            from app.ai_coach import recent_heat_stress
+
+            profile = (
+                db.query(AthleteProfile)
+                .filter(AthleteProfile.user_id == current_user.id)
+                .first()
+            )
+            heat_stress = recent_heat_stress(db, date.today(), current_user.id)
+            guidance = nutrition_mod.compute_fuelling_guidance(
+                duration_sec=float(duration_sec),
+                intensity="race",
+                weight_kg=profile.weight_kg if profile else None,
+                heat_stress=heat_stress,
+            )
+            if guidance:
+                resp.fuelling_guidance = FuellingGuidance(**guidance.__dict__)
+
+    return resp
